# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
    RegistryProject:
      type: object
      required: [id, tenant_id, name, public, created_at]
      properties:
        id:        { type: string }
        tenant_id: { type: string }
        name:      { type: string }
        public:    { type: boolean }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
    CreateRegistryProjectRequest:
      type: object
      required: [idempotency_key, name]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        name:            { type: string, minLength: 1, maxLength: 128 }
        public:          { type: boolean, default: false }
    RegistryProjectListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RegistryProject' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    RegistryRepository:
      type: object
      required: [project, name, artifact_count, pull_count]
      properties:
        project:        { type: string }
        name:           { type: string }
        artifact_count: { type: integer }
        pull_count:     { type: integer }
        permission:
          $ref: '#/components/schemas/RegistryPermission'
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    RegistryRepositoryListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RegistryRepository' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    RegistryArtifact:
      type: object
      required: [project, repository, digest, tags, media_type, size_bytes]
      properties:
        project:    { type: string }
        repository: { type: string }
        digest:     { type: string }
        tags:
          type: array
          items: { type: string }
        media_type:  { type: string }
        size_bytes:  { type: integer }
        pushed_at:   { type: string, format: date-time }
        scan_status:
          $ref: '#/components/schemas/RegistryScanResult'
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    RegistryArtifactListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RegistryArtifact' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    RegistryPermission:
      type: object
      required: [project, repository, subject, actions, state, updated_at]
      properties:
        project:    { type: string }
        repository: { type: string }
        subject:    { type: string }
        actions:
          type: array
          items: { type: string, enum: [pull, push, delete, scan] }
        state: { type: string, enum: [active, duplicate] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        updated_at: { type: string, format: date-time }
    SetRegistryPermissionRequest:
      type: object
      required: [idempotency_key, subject, actions]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        subject:         { type: string, minLength: 1, maxLength: 128 }
        actions:
          type: array
          minItems: 1
          items: { type: string, enum: [pull, push, delete, scan] }
    RegistryScanResult:
      type: object
      required: [image, status, critical, high, medium, low]
      properties:
        image:       { type: string }
        status:      { type: string, enum: [not_scanned, pending, running, complete, failed] }
        critical:    { type: integer }
        high:        { type: integer }
        medium:      { type: integer }
        low:         { type: integer }
        report_url:  { type: string }
        provider_id: { type: string }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        scanned_at: { type: string, format: date-time }
    CreateRegistryPullSecretRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        name:            { type: string, minLength: 1, maxLength: 128, default: ani-registry-pull }
        namespace:       { type: string, minLength: 1, maxLength: 128 }
    RegistryPullSecret:
      type: object
      required: [project, name, secret_ref, registry, username, created_at]
      properties:
        project:    { type: string }
        name:       { type: string }
        secret_ref: { type: string }
        registry:   { type: string }
        username:   { type: string }
        namespace:  { type: string }
        state:      { type: string, enum: [active, duplicate] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
    RegistryProjectScanReport:
      type: object
      required: [project, status, critical, high, medium, low, artifacts_total, scanned_artifacts]
      properties:
        project:           { type: string }
        status:            { type: string, enum: [not_scanned, pending, running, complete, failed] }
        critical:          { type: integer }
        high:              { type: integer }
        medium:            { type: integer }
        low:               { type: integer }
        artifacts_total:   { type: integer }
        scanned_artifacts: { type: integer }
        provider_id:       { type: string }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        scanned_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: 未认证或 Token 无效（code=UNAUTHORIZED）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问该资源（code=FORBIDDEN）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /registry/projects:
    get:
      operationId: listRegistryProjects
      summary: 列出镜像仓库项目
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:read"
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 镜像仓库项目列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryProjectListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createRegistryProject
      summary: 创建或确保镜像仓库项目
      description: 创建租户镜像仓库项目；客户端重试必须复用同一个 idempotency_key。
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:write"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRegistryProjectRequest' }
      responses:
        "201":
          description: 镜像仓库项目已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryProject' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/projects/{project}/repositories:
    get:
      operationId: listRegistryRepositories
      summary: 列出项目下镜像仓库
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:read"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 镜像仓库列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryRepositoryListResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/projects/{project}/repositories/{repository}/artifacts:
    get:
      operationId: listRegistryArtifacts
      summary: 列出镜像 artifacts
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:read"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: repository, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 镜像 artifact 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryArtifactListResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/projects/{project}/repositories/{repository}/permissions:
    post:
      operationId: setRegistryRepositoryPermission
      summary: 设置镜像仓库权限
      description: 为租户项目下的镜像仓库设置 pull/push/delete/scan 权限；客户端重试必须复用同一个 idempotency_key。
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:write"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: repository, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetRegistryPermissionRequest' }
      responses:
        "200":
          description: 镜像仓库权限已设置
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryPermission' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/projects/{project}/pull-secret:
    post:
      operationId: createRegistryProjectPullSecret
      summary: 创建镜像仓库拉取凭据引用
      description: 为项目创建 pull secret 引用；local profile 不返回真实密钥明文。
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:write"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRegistryPullSecretRequest' }
      responses:
        "201":
          description: Pull secret 引用已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryPullSecret' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/projects/{project}/scan-report:
    get:
      operationId: getRegistryProjectScanReport
      summary: 获取项目镜像扫描汇总
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:read"
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 项目扫描汇总
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryProjectScanReport' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /registry/images/scan-result:
    get:
      operationId: getRegistryImageScanResult
      summary: 获取镜像安全扫描结果
      tags: [Registry]
      x-ani-rbac-scope: "scope:registry:read"
      parameters:
        - { name: image, in: query, required: true, schema: { type: string, minLength: 1 } }
      responses:
        "200":
          description: 镜像安全扫描结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistryScanResult' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── 平台品牌化 ──────────────────────────────────────────────────────────────
...